+++ src/main/java/hanium/dtc/openai/controller/OpenAiController.java
@@ -1,8 +1,8 @@
-package hanium.dtc.controller;
+package hanium.dtc.openai.controller;
 
-import hanium.dtc.dto.openai.request.UserQuestionRequest;
+import hanium.dtc.openai.dto.request.UserQuestionRequest;
 import hanium.dtc.global.ResponseDto;
-import hanium.dtc.service.OpenAiService;
+import hanium.dtc.openai.service.OpenAiService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
