+++ src/main/java/hanium/dtc/openai/dto/request/UserQuestionRequest.java
@@ -1,4 +1,4 @@
-package hanium.dtc.dto.openai.request;
+package hanium.dtc.openai.dto.request;
 
 public record UserQuestionRequest (
         String question
