+++ src/main/java/hanium/dtc/openai/dto/response/OpenAiResponse.java
@@ -1,4 +1,4 @@
-package hanium.dtc.dto.openai.response;
+package hanium.dtc.openai.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
