+++ src/main/java/hanium/dtc/openai/service/OpenAiService.java
@@ -1,13 +1,12 @@
-package hanium.dtc.service;
+package hanium.dtc.openai.service;
 
 import hanium.dtc.config.OpenAiPrompt;
-import hanium.dtc.dto.openai.request.OpenAiRequest;
-import hanium.dtc.dto.openai.response.OpenAiResponse;
+import hanium.dtc.openai.dto.request.OpenAiRequest;
+import hanium.dtc.openai.dto.response.OpenAiResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestClient;
 
 @Slf4j
